Split db.integration.test.ts into per-domain test files
Split the monolithic 1,902-line integration test file into 7 focused
test files matching the db/ domain module structure:
- tags.integration.test.ts
- time-series.integration.test.ts
- activities.integration.test.ts
- health-connect.integration.test.ts
- mcp-sessions.integration.test.ts
- settings.integration.test.ts
- lastfm-rules.integration.test.ts

All 640 tests pass.
