+++ apps/backend/src/db/health-connect.integration.test.ts
@@ -1,0 +1,133 @@
+import { afterAll, beforeAll, beforeEach, describe, expect, test, vi } from 'vitest'
+import { cleanTestDb, getTestUser, startTestDb, stopTestDb } from '../test/db-test-helper'
+import { getDailyAggregateValue, processDailyAggregate } from './health-connect'
+
+const CONTAINER_TIMEOUT = 60_000
+
+describe('Health Connect Integration Tests', () => {
+  beforeAll(async () => {
+    await startTestDb()
+  }, CONTAINER_TIMEOUT)
+
+  afterAll(async () => {
+    await stopTestDb()
+  })
+
+  beforeEach(async () => {
+    await cleanTestDb()
+  })
+
+  describe('processDailyAggregate', () => {
+    test('stores valid steps aggregate', async () => {
+      const user = getTestUser()
+
+      await processDailyAggregate(user, {
+        dataOrigins: ['com.oura.ring'],
+        date: '2024-01-15',
+        metric: 'steps',
+        value: 10000,
+      })
+
+      const result = await getDailyAggregateValue(user, 'steps', new Date('2024-01-15'))
+      expect(result).toBe(10000)
+    })
+
+    test('rejects invalid metric with warning', async () => {
+      const user = getTestUser()
+      const consoleSpy = vi.spyOn(console, 'warn').mockImplementation(() => {})
+
+      await processDailyAggregate(user, {
+        dataOrigins: [],
+        date: '2024-01-15',
+        metric: 'invalid_metric',
+        value: 100,
+      })
+
+      expect(consoleSpy).toHaveBeenCalledWith(expect.stringContaining('Invalid metric in daily aggregate'))
+      consoleSpy.mockRestore()
+
+      // Should not have stored anything
+      const result = await getDailyAggregateValue(user, 'steps', new Date('2024-01-15'))
+      expect(result).toBeNull()
+    })
+
+    test('rejects non-cumulative metric with warning', async () => {
+      const user = getTestUser()
+      const consoleSpy = vi.spyOn(console, 'warn').mockImplementation(() => {})
+
+      await processDailyAggregate(user, {
+        dataOrigins: [],
+        date: '2024-01-15',
+        metric: 'heart_rate', // Valid but not cumulative
+        value: 72,
+      })
+
+      expect(consoleSpy).toHaveBeenCalledWith(expect.stringContaining('is not a cumulative metric'))
+      consoleSpy.mockRestore()
+    })
+
+    test('accepts all cumulative metrics', async () => {
+      const user = getTestUser()
+      const cumulativeMetrics = ['steps', 'distance', 'floors_climbed', 'calories_active', 'calories_total']
+
+      for (const metric of cumulativeMetrics) {
+        await processDailyAggregate(user, {
+          dataOrigins: [],
+          date: '2024-01-15',
+          metric,
+          value: 100,
+        })
+      }
+
+      // Verify all were stored
+      for (const metric of cumulativeMetrics) {
+        const result = await getDailyAggregateValue(user, metric as 'steps', new Date('2024-01-15'))
+        expect(result).toBe(100)
+      }
+    })
+
+    test('upserts on same day', async () => {
+      const user = getTestUser()
+
+      await processDailyAggregate(user, {
+        dataOrigins: [],
+        date: '2024-01-15',
+        metric: 'steps',
+        value: 5000,
+      })
+
+      await processDailyAggregate(user, {
+        dataOrigins: [],
+        date: '2024-01-15',
+        metric: 'steps',
+        value: 10000,
+      })
+
+      const result = await getDailyAggregateValue(user, 'steps', new Date('2024-01-15'))
+      expect(result).toBe(10000)
+    })
+  })
+
+  describe('getDailyAggregateValue', () => {
+    test('returns value when aggregate exists', async () => {
+      const user = getTestUser()
+
+      await processDailyAggregate(user, {
+        dataOrigins: [],
+        date: '2024-01-15',
+        metric: 'steps',
+        value: 12500,
+      })
+
+      const result = await getDailyAggregateValue(user, 'steps', new Date('2024-01-15'))
+      expect(result).toBe(12500)
+    })
+
+    test('returns null when no aggregate exists', async () => {
+      const user = getTestUser()
+
+      const result = await getDailyAggregateValue(user, 'steps', new Date('2024-01-15'))
+      expect(result).toBeNull()
+    })
+  })
+})
